Move from State to Store namespace
