+++ src/Vlingo.Symbio/Vlingo.Symbio.csproj
@@ -36,5 +36,8 @@
       <PackageReference Include="Vlingo.Actors" Version="0.4.8" />
       <PackageReference Include="Vlingo.Common" Version="0.5.6" />
     </ItemGroup>
+    <ItemGroup>
+      <Folder Include="Store\State" />
+    </ItemGroup>
 
 </Project>
